+++ src/main/java/com/b1n_ry/yigd/DeathHandler.java
@@ -62,10 +62,12 @@ public class DeathHandler {
         GraveComponent graveComponent = new GraveComponent(player.getGameProfile(), inventoryComponent, expComponent,
                 world, graveGenerationPos.add(0D, .5D, 0D), new TranslatableDeathMessage(deathSource, player), killerId);  // Will keep track of player grave (if enabled)
 
-        if (graveComponent.isEmpty() && respawnComponent.isEmpty()) return;  // There is literally no information worth saving
-
         GameProfile profile = player.getGameProfile();
-        graveComponent.backUp();
+        if (!graveComponent.isEmpty() || !respawnComponent.isEmpty()) {  // There is literally no information worth saving
+            Yigd.LOGGER.info("Did not backup data (grave data empty)");
+            graveComponent.backUp();
+        }
+
         respawnComponent.primeForRespawn(profile);
 
         Direction playerDirection = player.getHorizontalFacing();