+++ src/main/java/com/b1n_ry/yigd/components/GraveComponent.java
@@ -326,9 +326,6 @@ public class GraveComponent {
                 .with(net.minecraft.state.property.Properties.HORIZONTAL_FACING, direction)
                 .with(Properties.WATERLOGGED, waterlogged);
 
-        respawnComponent.setGraveGenerated(true);  // Not guaranteed yet, but only errors can stop it from generating after this point
-        DeathInfoManager.INSTANCE.markDirty();  // Make sure respawn component is updated
-
         // At this point is where the END_OF_TICK would be implemented, unless it wasn't already so
         Yigd.END_OF_TICK.add(() -> {
             BlockState previousState = world.getBlockState(pos);
@@ -337,13 +334,17 @@ public class GraveComponent {
             BlockPos placedPos = this.getPos();
 
             if (!placed) {
-                Yigd.LOGGER.error("Failed to generate grave at X: %d, Y: %d, Z: %d, %s".formatted(
+                Yigd.LOGGER.error("Failed to generate grave at X: %d, Y: %d, Z: %d, %s. Grave block placement failed".formatted(
                         placedPos.getX(), placedPos.getY(), placedPos.getZ(), world.getRegistryKey().getValue()));
                 Yigd.LOGGER.info("Dropping items on ground instead of in grave");
+                context.player().sendMessage(Text.translatable("text.yigd.message.grave_generation_error"));
                 this.getInventoryComponent().dropGraveItems(world, Vec3d.of(placedPos));
                 this.getExpComponent().dropAll(world, Vec3d.of(placedPos));
                 return;
             }
+
+            respawnComponent.setGraveGenerated(true);  // Not guaranteed yet, but only errors can stop it from generating after this point
+            DeathInfoManager.INSTANCE.markDirty();  // Make sure respawn component is updated
 
             GraveBlockEntity be = (GraveBlockEntity) world.getBlockEntity(placedPos);
             if (be == null) return;